requirements.txt
boto3
python-dotenv
faker
mysql-connector-python